use usize for pointer values

+++ runtime-interface/src/sandbox_mod/detail/get_instance_ptr.rs
@@ -28,12 +28,11 @@ pub fn method(self_: &mut dyn FunctionContext, instance_id: u32) -> HostPointer 
         method_result = SANDBOXES.with(|sandboxes| {
             let instance = sandboxes
                 .borrow_mut()
-                .get(data_ptr as u64)
+                .get(data_ptr as usize)
                 .instance(instance_id)
                 .expect("Failed to get sandboxed instance");
 
-            instance.as_ref().get_ref()
-                as *const gear_sandbox_host::sandbox::SandboxInstance
+            instance.as_ref().get_ref() as *const gear_sandbox_host::sandbox::SandboxInstance
                 as HostPointer
         });
     });
