+++ runtime-interface/src/sandbox_mod/detail/set_global_val.rs
@@ -30,7 +30,7 @@ pub fn method(self_: &mut dyn FunctionContext, instance_idx: u32, name: &str, va
         let result = SANDBOXES.with(|sandboxes| {
             let instance = sandboxes
                 .borrow_mut()
-                .get(data_ptr as u64)
+                .get(data_ptr as usize)
                 .instance(instance_idx)
                 .map_err(|e| e.to_string())
                 .expect("Failed to set global in sandbox");
